--- src/blocks/index.ts
@@ -1,2 +1,0 @@
-export * from "./ezpr";
-export * from "./help";
