+++ src/cmd/index.ts
@@ -1,2 +1,4 @@
-export * from "./ezpr";
-export * from "./help";
+export * from "../ezpr/cmd";
+export * from "../help/cmd";
+export * from "./getNameByID";
+export * from "./modal";
