--- src/cmd/interface.ts
@@ -1,4 +1,0 @@
-export interface ICommand {
-  input?: string;
-  handle(): any;
-}
