+++ src/ezpr/cmd.ts
@@ -1,10 +1,9 @@
-import { Block, KnownBlock, SayFn } from "@slack/bolt";
+import { Block, KnownBlock } from "@slack/bolt";
 import { View, WebClient } from "@slack/web-api";
-import { ezprMessage } from "../../blocks";
-import { EZPRArguments } from "../../types";
-import { ICommand } from "../interface";
-import { OpenModalCommand } from "../modal";
-import ezprModal from "../../blocks/ezpr/modal.json";
+import { ezprMessage } from "./blocks";
+import { EZPRArguments, ICommand } from "../types";
+import { OpenModalCommand } from "../cmd/modal";
+import ezprModal from "./modal.json";
 
 export class EZPRCommand implements ICommand {
   client: WebClient;
