+++ src/help/index.ts
@@ -1,0 +1,4 @@
+export * from "./";
+export * from "./blocks";
+export * from "./cmd";
+export * from "./slash_command";
