+++ src/help/slash_command.ts
@@ -1,5 +1,5 @@
 import { SlashCommand } from "@slack/bolt";
-import { HelpArguments } from "../../types";
+import { HelpArguments } from "../types";
 
 export function ParseSlashHelpCommand(payload: SlashCommand): HelpArguments {
   const input = `${payload.command} ${payload.text}`;
