--- src/parse/ezpr/index.ts
@@ -1,2 +1,0 @@
-export * from "./form_submission";
-export * from "./slash_command";
