--- src/parse/help/index.ts
@@ -1,1 +1,0 @@
-export * from "./slash_command";
