+++ src/parse/index.ts
@@ -6,8 +6,9 @@ import {
 } from "../constants";
 import { StringDictionary } from "../types";
 
-export * from "./ezpr";
-export * from "./help";
+export * from "../ezpr/form_submission";
+export * from "../ezpr/slash_command";
+export * from "../help/slash_command";
 
 const EMPTY_STRING = "";
 const SINGLE_QUOTES = "'";
